xDAI/Rinkeby support (#1)
Adds support for xDAI and Rinkeby solutions as well. Mainly adding the different solver addresses, graph endpoints and etherscan URLs for different networks as well as adding a radio selector to the top of the list.

<img width="1295" alt="Screen Shot 2020-10-14 at 13 00 10" src="https://user-images.githubusercontent.com/1200333/95993123-fe242c00-0e2e-11eb-85e9-8488a00a6f98.png">


### Test Plan

Manual testing. At the moment I cannot check if the solutions file are readible, due to some CORS issue with the actual AWS bucket (fetching the list of files fails). I asked DevOps to change the CORS policy and will test before merging.